Extract backup logic into new file
